move prompt inputs to the ask package
