+++ cli/pkg/tui/ask/errors.go
@@ -1,4 +1,4 @@
-package tui
+package ask
 
 import "errors"
