+++ cli/pkg/tui/ask/select.go
@@ -1,4 +1,4 @@
-package tui
+package ask
 
 import (
 	"fmt"
