+++ cli/pkg/tui/ask/textinput.go
@@ -1,4 +1,4 @@
-package tui
+package ask
 
 import (
 	"strings"
