+++ cli/pkg/tui/ask/toggle.go
@@ -1,4 +1,4 @@
-package tui
+package ask
 
 import (
 	"fmt"
